Factor out logging (#1262)
